Backend/requirements.txt
fastapi==0.114.0
uvicorn[standard]==0.30.0
python-dotenv==1.0.1
httpx==0.27.0
pydantic==2.8.2
# AI & ML
openai==1.3.0
langchain==0.0.340
langchain-community==0.0.10
faiss-cpu==1.7.4
pinecone-client==2.2.4

# Vision & Images
pillow==10.0.0
opencv-python==4.8.1
pytesseract==0.3.10

# Math & Science
sympy==1.12
numpy==1.24.3

# Video & Audio
moviepy==1.0.3
manim==0.17.3
elevenlabs==0.2.0

# Audio Processing
speechrecognition==3.10.0
pydub==0.25.1

# PDF Processing
pypdf==3.15.0
pdf2image==1.16.3       